refactor(tools): extract registerSearchTools into tools/search.js
Moves 12 read-only tools (find_scenes, get_scene_prose, get_chapter_prose,
get_arc, list_characters, get_character_sheet, list_places, get_place_sheet,
search_metadata, list_threads, get_thread_arc, get_relationship_arc) out of
index.js and into tools/search.js via registerSearchTools(s, ctx).

Extends toolContext with the six values these tools need: GIT_ENABLED,
paginateRows, DEFAULT_METADATA_PAGE_SIZE, MAX_CHAPTER_SCENES,
getSceneProseAtCommit, readSupportingNotesForEntity, readEntityMetadata.

Write tools interspersed in the same region (create_character_sheet,
create_place_sheet, upsert_thread_link) stay in index.js for the
upcoming metadata group PR.

178 unit + 137 integration tests pass. Lint clean.
